+++ src/qml/qml/qqmltype_p.h
@@ -199,11 +199,11 @@ public:
 
 private:
     friend class QQmlTypePrivate;
-    friend uint qHash(const QQmlType &t, uint seed);
+    friend size_t qHash(const QQmlType &t, uint seed);
     QQmlRefPointer<const QQmlTypePrivate> d;
 };
 
-inline uint qHash(const QQmlType &t, uint seed = 0)
+inline size_t qHash(const QQmlType &t, uint seed = 0)
 {
     return qHash(reinterpret_cast<quintptr>(t.d.data()), seed);
 }
