+++ src/quick/scenegraph/coreapi/qsgbatchrenderer.cpp
@@ -4637,7 +4637,7 @@ bool operator!=(const GraphicsState &a, const GraphicsState &b) Q_DECL_NOTHROW
     return !(a == b);
 }
 
-uint qHash(const GraphicsState &s, uint seed) Q_DECL_NOTHROW
+size_t qHash(const GraphicsState &s, uint seed) Q_DECL_NOTHROW
 {
     // do not bother with all fields
     return seed
@@ -4665,7 +4665,7 @@ bool operator!=(const GraphicsPipelineStateKey &a, const GraphicsPipelineStateKe
     return !(a == b);
 }
 
-uint qHash(const GraphicsPipelineStateKey &k, uint seed) Q_DECL_NOTHROW
+size_t qHash(const GraphicsPipelineStateKey &k, uint seed) Q_DECL_NOTHROW
 {
     // no srb and rp included due to their special comparison semantics and lack of hash keys
     return qHash(k.state, seed) + qHash(k.sms->programRhi.program, seed);
