+++ src/quick/scenegraph/coreapi/qsgbatchrenderer_p.h
@@ -645,7 +645,7 @@ struct GraphicsState
 
 bool operator==(const GraphicsState &a, const GraphicsState &b) Q_DECL_NOTHROW;
 bool operator!=(const GraphicsState &a, const GraphicsState &b) Q_DECL_NOTHROW;
-uint qHash(const GraphicsState &s, uint seed = 0) Q_DECL_NOTHROW;
+size_t qHash(const GraphicsState &s, uint seed = 0) Q_DECL_NOTHROW;
 
 struct ShaderManagerShader;
 
@@ -659,7 +659,7 @@ struct GraphicsPipelineStateKey
 
 bool operator==(const GraphicsPipelineStateKey &a, const GraphicsPipelineStateKey &b) Q_DECL_NOTHROW;
 bool operator!=(const GraphicsPipelineStateKey &a, const GraphicsPipelineStateKey &b) Q_DECL_NOTHROW;
-uint qHash(const GraphicsPipelineStateKey &k, uint seed = 0) Q_DECL_NOTHROW;
+size_t qHash(const GraphicsPipelineStateKey &k, uint seed = 0) Q_DECL_NOTHROW;
 
 struct ShaderManagerShader
 {
