+++ src/quick/scenegraph/coreapi/qsgtexture.cpp
@@ -83,7 +83,7 @@ bool operator!=(const QSGSamplerDescription &a, const QSGSamplerDescription &b) 
     return !(a == b);
 }
 
-uint qHash(const QSGSamplerDescription &s, uint seed) Q_DECL_NOTHROW
+size_t qHash(const QSGSamplerDescription &s, uint seed) Q_DECL_NOTHROW
 {
     const int f = s.filtering;
     const int m = s.mipmapFiltering;
